Fix double nesting of sections json
